+++ backend/functions/src/presentation/routes/index.js
@@ -1,0 +1,20 @@
+const { Router } = require('express');
+const router = Router();
+
+const { notFound } = require('./notFound');
+const { errorHandler } = require('./errorHandler');
+
+const welcome = {
+  title: 'Welcome to MTG Deals',
+  message: 'If this is your first time? <br /> <b>Enjoy it!</b>',
+};
+
+router.get('/', (req, res) => res.send(process.env.TEST_SECRET || 'Failed'));
+router.get('/welcome', (req, res) => res.send(welcome));
+router.get('/:userId/welcome', (req, res) => res.send(welcome));
+
+// These two MUST be the last ones
+router.use(notFound);
+router.use(errorHandler);
+
+module.exports = { router };
